golang: fix support for go modules in subdirectories
This commit fixes situations where a Go module is in a subdirectory of a
repository, as might often be the case in monorepos.

Previously, commands were attempted to be executed at the root of the
repo, which would fail because with Go modules, we can no longer rely on
a global GOPATH to tell us where the code lives, commands are only valid
when executed from within the Go module itself.