improve error messages when running commands (#510)
* improve error message when running command

When running a command, it's helpful to include the command that
was executed in the error. This makes it easier for the user to drill
down and diagnose problems.

* fix stderr capturing

This fixes a bug where the stderr of a failed command was never captured,
and thus, diagnosing problems when a particular command failed become much
harder.

The specific issue here was that a stack copy of a `bytes.Buffer` was
returned, while handing a pointer to it to the `*exec.Cmd`. Writes to
the buffer throgh `*exec.Cmd` would consequently not update the stack
copy returned.

A regression test is added to ensure that stderr is captured.