Add higher-level interface for graph building; fix golang transitive analysis (#10)
This adds a higher-level interface for folding data structures into graphs, and uses the new interface to resolve a bunch of issues faced with the golang analyzers. In particular: missing edges and nodes in the transitive dependency graph are now correctly recorded.

The higher-level interface also sets the stage for combining several analysis strategies. Something we've noticed when building analyzers is that the "better" strategies often give less information about edges or tags. This new interface will make it easier to include metadata from worse strategies.