Use `cabal v2-build all` for generating `plan.json`
This commit modifies the command for generating the `plan.json` file
for Cabal projects to use the target `all`. Running `v2-build` without
a target, as was done previously, only works for projects that have a
package in their top-level directory. Using a target `all` works for
all projects and behaves identically as before for projects with a
top-level package.