FFI WIP
It works in C but doesn't work in Haskell. Why? Who knows??

When the Haskell build flags were wrong, it generated an error
about the actual library missing, which isn't the case
anymore.
This seems to imply that the library is configured correctly
but for some reason Haskell can't see the actual symbols.

I'm timeboxing this for now.