fix(ANE-2243): use correct Docker Hub registry API endpoint (#1500)
* fix(ANE-2243): use correct Docker Hub registry API endpoint

When analyzing Docker images from Docker Hub, we now correctly use registry-1.docker.io
instead of index.docker.io as the API endpoint. This fixes a 403 error that occurred
when trying to analyze Docker images.

The change distinguishes between:
- index.docker.io: Used for parsing Docker image references
- registry-1.docker.io: Used for making actual API requests

Jira: ANE-2243

* Revert allow-newer additions to cabal.project

* Update Changelog.md

* fix: remove unused import

* chore: ignore hoogle.haskell.org in link checker

The link checker sometimes fails on hoogle.haskell.org due to rate limiting
or other issues, but the link is known to be valid. Add it to the ignored
patterns.

* Revert .markdown-link-check.json

---------

Co-authored-by: ryan link <ryanlink@ryans-mbp.lan>